requirements.txt
langchain-openai
langchain
langchain-core
openai
geopy
python-dotenv
googlemaps